sped up the borders and made stencil for points adding speeding things up
